Adding better text patching to adjust alignment and font name
Previously we simply overwrote all FontName props to SansSerif to prevent text from
being outlined. We continue to do that but cache the original FontName to restore
it. We also change the HorizontalAlignment and VerticalAlignment to left, top in
order to ensure that the text is in the correct place when the font is later
corrected in the SVG file.

We also update the inkscape PDF conversion call to use the --export-filename flag instead
of --export-pdf, which was removed.